Don't open settings on every connect with no SDK.
This logic spams the settings dialog on join workspace.
It breaks when there a stomp prompt appears.
I also removed the warning about an SDK because some forks of intellij
do not have the concept of an SDK.
The user just needs to know how to use their editor. If they don't and
they have problems we can ask them to check their IDEA.log. We'll see
the SDK warning in it.